Add DateTimehelper functions
