+++ app/build.gradle.kts
@@ -68,6 +68,7 @@ dependencies {
     implementation(projects.feature.sydneyTrains.database.real)
     implementation(projects.feature.sydneyTrains.network.api)
     implementation(projects.feature.sydneyTrains.network.real)
+    implementation(projects.feature.tripPlanner.domain) // TODO not really required remove it later
     implementation(projects.feature.tripPlanner.network.api)
     implementation(projects.feature.tripPlanner.network.real)
